Fix compiler warnings.

+++ source/mango/image/image_exr.cpp
@@ -2121,10 +2121,10 @@ const u8* ContextEXR::decompress_b44(Memory dest, ConstMemory source, int width,
 
         for (const Channel& channel : channels)
         {
-            size_t nx = div_ceil(block_width, channel.xsamples);
-            size_t ny = div_ceil(block_height, channel.ysamples);
-            size_t bpl    = nx * channel.bytes;
-            size_t nBytes = ny * bpl;
+            int nx = div_ceil(block_width, channel.xsamples);
+            int ny = div_ceil(block_height, channel.ysamples);
+            int bpl = nx * channel.bytes;
+            int nBytes = ny * bpl;
 
             if (nBytes == 0)
                 continue;
